requirements.txt
streamlit
python-dotenv
groq
sounddevice
scipy
soundfile